Use set for Relation body, various cleanups.
Every row of a relation must be unique.  The easiest way to guarantee
this is to use a set.  This requires that everything stored in a
Relation be hashable, but that's OK because everything we store in our
relational database is required to be constant data anyway.  However, it
does mean adding appropriate __hash__ methods to the Scaler and Row
classes.

I also realized that the 'rows' attribute was in the relation attribute
name space, and needed to be renamed _rows_ (since it is part of the
public API).  I considered making _sorted_attrs_ a private variable, but
the list of attribute names is needed as part of the public API, so I
instead renamed it _attr_names_, which is more descriptive.  The public
API gives no ordering promise, but internally we know it is sorted and
use it that way.

There are also a few other small cleanups, like deleting a commented out
method that is no longer used.