Add structure to test file, clarify several things.
The test file is clearly becoming more than just a literate test file.
So I added some subheadings to further structure the discussion.  I also
clarified some statements, and correct the one about Tutorial D's
handling of types and dinsd's presumed compliance (or rather lack there
of) to the type requirements of TTM.

I added a discussion of the naming convention (relation attributes
no leading '_', public API methods '_xxx_'), which caused me to
do the review and corrections in the previous checkin.

Finally, this checkin contains the start of the next chapter, on
relational operators, with the declaration of the Relations we'll be
using in the examples.