Really fix the Relation __eq__, and all compares.
I completely confused myself about what the __eq__ method was doing.
The correct fix is *much* simpler: just compare the row sets.  Note that
this works because all the values stored in the relation are the typed
values, so if all the values match all their types must match and the
relations must have the same type.  But this also means that I had to
get compare working correctly for all the values, which it turns out
they weren't.  While I was at it I extended the rich compare to all the
classes that need it.

(The test failure turned up randomly, since the tuples
would often be returned in the same order and thus the test would pass.)

I've also added __iter__ and __len__ methods to Relation, which we
will need shortly.  I considered turning Relation into a set
subclass, but I don't think I want it to have *all* of the methods
of set, just most of them.