Add intersect and times.
This required fixing join to support there being no attributes in
common.  I also promoted _header to a "public" API, _header_, to make it
easier to compare headers.  And I simplified the prototype for join.  I
had been thinking that 'first' and 'second' as explicit arguments would
give better error messages, but it turns out join(x) is supposed to be
legal (as is join(), but I'm not dealing with that yet).