Add project operator and syntax.
For better or worse I've allowed projection to be done via '<<' and
'>>', since Tutorial D has a simplified syntax for projection.

These examples also revealed a bug in display, where it would not handle
printing a single column due to a quirk in the way attrgetter treats its
arguments.  I filled an enhancement request bug (#16457) for providing a
way to get attrgetter and itemgetter to always return a tuple.