Dum and Dee, and related fixes.
Now we get to deal with join(), as well as projection to {}.  Getting
this to work required tweaking the display function, and allowing the
relational operators to accept zero arguments.  Project also needed
fixes to handle empty sets in both directions.  I'm not 100% sure "rel
>> {} == Dee" is correct (AITDT doesn't say, when it introduces them),
but that is what fell out of the code and I suspect it is correct.

I was impressed in general by how much just fell out correctly in the
code.  In particular I was surprised that DumDee({}) produced Dee
without any code changes.