Exercise 2 from Chapter 4, message fixes.
In addition to finding a way to improve the error messages that result
from type errors when creating a relation, I discovered a problem with
the class based method of defining type functions.  If you use a
function, it turns into a bound method and gets passed self, but only
when called via the instance...and I'm not at all sure why that happens.
There may well be a way to fix it, but I just worked around it in the
example for now.