Refactor Relation attributes into header attr.
We don't need to access the attribute names on the relation object, so
all that having them in the top level namespace did was make the method
names weird.  And we needed a dict of the attributes anyway.  So this
refactor removes the attributes from the class dictionary and just keeps
them in the header dict, now named just 'header', renaming the other
Relation methods that were _xxx_ to just xxx.

My test file caught all of the renames, which means my test coverage is
pretty good.

Oh, and I've decided that the ability of display to only show selected
columns is not only an attractive nuisance (a couple of my examples
produced incorrect tables and I didn't notice) but also can produce an
image that is *invalid* as a relational table, which is pedagogically
unsound.  So now you have to specify all of the columns.  I *might*
relax that later to allow you to specify just some column names but
display them all, but I don't have a need for that yet.