More examples, and clarify an error message.
By adding the parent relation name to the RowClass name the
attribute error message is much easier to trace back to its
source.  In fact, since the attribute names are part of the
name of a synthetic relation, it was immediately obvious
what typo I had made.

Could get a bit clumsy for tables with lots of attributes,
though.  Might not be sustainable.