Refactor 'all_but' as wrapper rather than keyword.
This is in aid of the upcoming 'group', which will also need the
projection style listing of attribute names, but for which I prefer not
to have an all_but keyword.  It is also a cleaner notation, with the
downside that it is somewhat more magical.

Oh, and in the process I discovered that I've been bending the Python
literal notation when it comes to empty attribute sets.  After
consideration I decided to preserve it, so I explained the issue in the
test doc so I don't forget about it when I do the real docs.