ungroup, union as aggregator, and group edge cases.
I dropped the questionable support for passing a type to union.  It
seems inconsistent with the rest of the system, and I don't know why
Tutorial D includes the ability to specify a type for an empty union.

Frankly, I think the dinsd formulation of the formal definition of
ungroup is clearer than the Tutorial D one.  But, then, being a
Python hacker I would think that :).

The edge cases for group are kind of fun (and not mentioned in the
book), and it is really cool that they just worked when I tried them.