Strings instead of lambdas for the win.
It occurred to me that is is *much* harder to have cross-python-platform
support for this library if I use bytecode inspection to go from
lambda expressions to SQL.  On the other hand, AST parsing a Python
expression is probably supported by most platforms, or we could write
a minimal parser of our own if needed.  Doing that parsing to turn
what we can into SQL will be *much* easier than bytecode parsing,
I'm pretty sure.

As an added bonus, writing expressions is now *much* simpler, and
the expression can refer directly to the attribute names and it will
do the right thing.

The only downside is the need for an 'expression namespace'.  And
that is actually an upside if you think about it: we are restricting
the set of things that can get involved in a computation to a known
set.  The tricky bit is managing that namespace...the global variable
solution adopted here is probably not what I'm going to end up with
long term, but I haven't figured out how I want to handle it yet.

Oh, and I was able to eliminate 'column'.  I also fixed a bug in
the repr of an empty relation, and added a test for it.