Relation comparison.
This turned out to be a much bigger deal that I expected.  This is
because I had to fix the fact that unlike relations could be compared
under my naive rich comparison scheme.  Putting a relation type check in
to the compare then revealed that we also needed type comparison support
in Row, since part of comparing relations is comparing their types, and
one of the possible attribute types is a Row.  Fixing that lead me down
a merry path involving weirdnesses in how ABCMeta classes work
(involving storing the type in a weakref dictionary, apparently...which
seemed to require the ability to compare types *before* they have been
initialized).  So, I wound up just ripping the Mapping subclass out,
since I'm not actually using it.  Had to fix a couple bugs that were
revealed by removing it, though, including letting Row act correctly
as a type function instead of re-validating all the data, and adding
a new case to Relation's __init__ giving a similar shortcut path
when the Row type of the input matches the Row type of the relation.

All in all, cleaner code.  And along the way I improved some more of the
error messages.