Add tests for Row errors without refactoring Row.
After due consideration and rereading the section of the book where
TUPLE is first discussed, I've decided not to factor out RowClass
into a public API.  I may change that later, but even if I do it
won't be instead, it will be in addition, for certain special
purposes.

In trying to write complete tests I found that if the first element
of the list is a dict, the constructor happily accepted Rows later
in the list (and ran them back through Row, an unneeded step).  So
I just made that part of the API and made it work in the other
direction as well (Row instance first in list).