Provide row as an API after all.
I'm beginning to understand how this is supposed to work.  So
I've defined a row function for defining the dinsd correspondence
to a TUPLE literal.  Next I'll do the RELATION literal, which
will cause me to completely rewrite the section of the testdoc
I just added row to.

Oh, and doing this revealed that I hadn't fixed the equality
comparison for Row, which I have now done.