{} call form for rel, and __name__ keyword.
We now have a rel form that looks more like Tutorial D's RELATION.  And
we have a form that *acts* more like Tutorial D's RELATION when defining
a named relation.  So you get to pick whether you prefer Python's loser
rules, or a slightly more awkward form with stricter rules.

I'm not 100% convinced the {} form for rel is a good idea.