row discussion, fail to make header and degree r/o.
In writing up row I realized that I needed to discuss '_' names as part
of the main text in the rel section, which lead to the "edge cases"
section getting completely integrated into the text.  I added a mention
of degree to the rel section as well.

And then I realized that degree depended on header, and that header
really ought to be read only.  I tried to make it so, but I'm getting
errors that header and degree are unknown attributes, and that is true
even if I duplicate the properties on the Relation class.  So there's
something I'm not understanding about metaclasses, but I'm going to
postpone dealing with that for the moment.

There's more to come in the row section.