Relation literals, some error message improvements.
Adding this feature required a few tweaks.  For some reason it also
revealed a bug: I was still using row for the special row variable in
string expressions, even though I've now adopted row for the name of the
row literal function.  So I changed that to r.

I altered the Relation repr to match the relation literal syntax that
matches Tutorial D most closely, since that presentation seems to me to
convey the maximum amount of information about the relation, including
the fact that the rows cannot have duplicate keys, and that the
rows of the relation are a set.  It is much wordier, but it is also
the repr so I think that's OK.

Debugging the new feature also resulted in some error message
improvements.