dict form for rows in rel, and row sorting fix.
After banging my head on the different ways to call rel and Relation
__init__, I think I've finally more or less sorted it out, and come up
with an understandable explanation and code scheme.  The tuple form is a
special exception and supported only in the multi-argument form of the
rel call.  Otherwise we all either a set of rows or an iterator of a mix
of dicts and rows, and these may be passed either as a multi-argument
list or as a single argument iterator.  The rational for allowing both
dicts and rows is that they really are logically equivalent.  The plan
is to support either wherever the other is allowed, and probably (as
indicated in the doc change) not mention row objects except in
passing in the ultimate tutorial.

Along the way I found a bug in the sorting of tuples, and simplified the
code for sorting in the relation repr.  Probably the new way is less
efficient, but that doesn't matter at this stage.