More rel tests, relation literal bug fix, msg fix.
The test for a single row literal that the text says "should be obvious
but needs to be tested" sure did, it revealed a bug.  I also improved
the duplicate row error message by adding the row number, and the
value/type error one by also catching ValueError.

Added some other tests, including Dum/Dee literal tests, and realized in
the process that to really conform with TTM (I checked) the type name
literally needs to be Relation({<headerdef>}).  Or, rather, what is
shown in the repr needs to be that.  That is, no named types like
IsEnrolledOn, meaning we ditch the class version of the declaration, as
well as the __name__= hack.

This will simplify the code slightly, and the conceptual framework
nicely.  Eventually I'll end up with a type cache and there will be only
one instance of each distinct type.  Which would allow me to drop the
__eq__ methods on the metatype, which would be nice.

(And yes, that is a legitimate use of the word "literally" there :).