Namespace context manager.
I've decided on an approach to the namespace management.

First, expression_namespace is confirmed as a permanent part of the API,
but only for registering global stuff (application defined functions and
types, principally).

Second, we have a thread local context manager, and anything in that
namespace when Row._as_locals_ is called gets added to locals.

Third, we will automatically add the names of the databases in the
current database to locals if the relation being operated on is
associated with a database (details to follow :).

I haven't tested the threading support, and won't until I get as far as
introducing database connections and threading, which won't be for quite
a while.

I also decided to hold my nose and import all other namespaces using a
leading '_', thereby avoiding repeating the name list.