Simplify metaclasses, make row header/degree r/o.
We don't need the __eq__ methods now that we can compare by
object identity, and we don't need _Relation's __new__
method since all creation goes through rel now.

Also simplified _get_type's call signature.

I thought about eliminating the metaclasses entirely based on Python's
consenting adults philosophy, but it is just too easy to *accidentally*
change the header through mutation of the dictionary if what is returned
is not a copy.  And the relation types are objects likely to be
manipulated by a dinsd application, so it is not enough just to make
the object instances do the safer thing.