Finish text refactoring of relation extension.
Also changed row str version from looking like a tuple of equates
to looking like a set of equates, which I think is the better
mnemonic representation.

Edge Test have moved inline instead of being in a separate section at
the end, since that was the style of the earlier sections.  Hopefully I
didn't miss any.