Finish aggregators rewrite, make avg correct on 0.
My attempt to make avg "better" by returning 0 for an empty relation is
mathematically incorrect and misguided, so I reverted it.  Ironically, I
have to explicitly do the division by zero as one leg of an if test,
since my loop variable is not defined if the loop doesn't iterate.
(That if test is what led to my mistake.)