Rewrite 'motivation' section as 'with' section.
This is the first introduction of with, and I use it several times, so
it seems reasonable to retitle the section even though "with" doesn't
have much to do with relation valued attributes.

I understand what the Tutorial D code is doing better now, so I've
refined my corresponding examples and comments.