Rewrite through ungroup, '_'ify special exp names.
I realized that blindly following Tutorial D into special syntax land
didn't really make sense, that having summarize really be just a
composition of compose and extend, with exactly the same semantics (the
thing passed to the expression is a row) would work just fine if the
"special name" for the summarized column was an attribute name.  This
lead to renaming it _summary_, and providing a back-door way to create
such a column in extend (_name_check=False).  As a bonus, the summarize
function became simpler.  I suppose I should consider rewriting it for
efficiency, but that is a premature optimization.

In doing ungroup, I realized that the special name "r" was a potential
collision with an attribute name, so I renamed it _row_, which makes me
much happier.

I also changed the error message when a project attribute isn't one of
the ones in the relation from "invalid" to "unknown", because having it
say invalid confused me for a good five minutes when I got that error on
_summary_.

Oh, and debugging my use of locals() revealed that I had been specifying
the name spaces in the wrong order in my eval calls.