row operators, compose as +.
I decided to make compose + so that I could avoid the issue of having
to have a row_compose function for the row operation.  Not perhaps the
best reason, but not a terrible one, I think.

I've implemented the row operations in terms of the relational
operators.  This is not perhaps efficient, but it certainly minimizes
duplicated code.  Later we may need to expand the tests for these, if
the non-pure-Python implementations don't use the same dodge (and
they may well not...creating a temporary in-db table just to extend
a row certainly seems like overkill).