Database objects and basic persistence.
The support for persistence using sqlite is very stupid and rather a
hack.  But it works, at least so far.

By 'stupid' I mean that we load all the tables into memory from the
persistent store when we open the database, and right now we rewrite all
of the rows when we do an update.  The only reason for using sqlite for
this rather that straight pickles is that when we introduce insert and
update we will only do single row insert/update for those, rather than
rewriting the entire database on each save, which we'd have to do for
pickle.

You will note that the above implies that this code is not thread safe
if there are multiple connections to the same database, and it cannot be.

I'm planning a smarter version of sqlite persistence, but it is a good
question how much code that will share with this implementation of
dinsd.  It may be that the only thing in common will be the API.
Assuming it works....