Save constraints to DB, rewrite reldef logic.
Since I have to add a table for the constraints, the cute trick of using
the type name to avoid having to create a relation definitions table is
no longer very meaningful.  So I dropped it in favor of the much more
reasonable reldefs table.

I really wanted to use a foreign key to the sqlite_master table, rather
than create that stupid single column _relnames table, but sqlite
doesn't support doing that.

As a result of using a keyword as a column name ('constraint'), I have
now properly quoted the identifiers in all of my sql.

Also added a note about the current execrable state of the exception
classes returned for various errors.