Key constraints.  But not completely implemented.
As I modified display, I realized that there needs to be a major
refactoring here, so that there is a subclass of _Relation for relations
stored in the db.  And display should be a method...and lots of other
related changes.

So there is one heck of a refactoring hanging off the "improve it" in
that XXX comment...