Major file refactor; Doc does not match code.
The tests pass, but I haven't touched the text at all (except for the
__all__ test, which I had to change), and there's a lot more refactoring
to do in the sqlite_pickle_db.py file.

We are now back to a package.  I have a non-trivial __init__ file (it
contains all of the algebra) because I don't want the names of the rel
and row classes to have an extra level in them.  This is perhaps a silly
reason, but it also makes logical sense to me: what is in the __init__
file is the "base API", and what is in the other modules is selectable
pieces of the extended API.