Best of both worlds: Database->dict, keeping r.
At first I was going to just drop the r attribute, because it was an
ugly hack.  Making the Database a dictionary was a no-brainer once I
thought of it (the existence of '_iter_rels' was a big clue).  When I
started rewriting the doc in terms of dictionary access, though, I was
sad.  So instead I converted r into *just* a relation-as-attribute
access method.  The whole thing is much cleaner now, I think.