PersitentRelation, and display as method.
This clears up that pesky little XXX comment in _display that
I ought to improve that if statement.  It also sets the stage
for a lot of what should be coming later.

Note that I have not yet done a sweep for places in the algebra or
docs where I'm using type(r)() and should be using rel(r.header)().

Also note that I don't know why the repr for a PersistentRelation class
is 'dinsd.PersistentRelation instead of
'sqlite_pickle_db.PersistentRelation', but I'm ignoring that question
for now.