Pointer to key on keyed relation.
This was a bit of a rabbit hole, as it revealed that I hadn't
thought about the fact that closing the database leaves the db
relation objects intact if there are references to them.  So now
I invalidate them on close by changing their __class__.

Also fixed up some wording per previous refactorings.