Use threading.local correctly in ns contextmanager.
I don't know what I was thinking when I implemented the ns context
manager, but I was very wrong.  Reimplemented it using threading.local
correctly (still no threading tests, though).  I also was thinking wrong
when I wrote that the namespace got used as globals; it doesn't, it is
part of locals, so we can indeed use a ChainMap.  We lose the attribute
notation for putting things in the namespace, but since namespaces
elsewhere in Python are dictionaries, this is probably actually a good
thing.

Not sure if I'm going to fix transactions next, which has the same
problem, or if I'll write threading tests.