Make db changes visible inside transaction (only).
This __getitem__ implementation feels a little hackish, but it works,
without having to employ locks.  Note that my example doesn't prove that
the persistent DB is actually updated.  I'm not sure how/where I want to
add those tests.

I also don't have a test of two transactions (as opposed to one
trans and one non-trans) not interfering, but I think I'll only add
such a test if there proves to be a bug there someday, since for any
implementations I've thought of this one should be the more important,
and sufficient, test.