'insert' method for db relations.
There will be a bunch of refactoring that will happen once these
mutation methods are available.  And a lot more tests are needed.

The big change here aside from adding insert is to move the
assurance of a transaction into a decorator, and simply requiring
that a transaction be active when the _check methods are run.