Add 'prototype' arg to extend.
It turns out that there are very good reasons to want to be able to
extend an empty relation.  This came up specifically in dealing with the
missing values problem (Missing-info-without-nulls.pdf), where there
might not be any missing values, but you still want to extend that table
when composing a view containing a column depicting both the real values
and the missing values.