Add some type guessing to extend.
I thought of a clever trick that will cover 80% or more of the cases
of extending empty relations: construct a dummy row by calling the type
functions with no arguments.  It might be worth adding this as
a further requirement on dinsd type functions.  The downside is that
it means you don't get type checking for null input values, so I'm not
going that route yet.