Allow types that do not except instances as input.
Datetime, which is going to be commonly used in tables, is an
example of such a type.  So that restriction needs to be removed.
This removes it from row, I'm not sure where else it might
need to be removed from.

Also add a test that ~ can't be used on an empty relation, since
that question came up and I didn't see the answer in my docs :).